% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{label_strategies}
\alias{label_strategies}
\title{Label strategies for volcano plots}
\format{
A list of functions:
\describe{
\item{candidate_pval}{Label top significant candidates by p-value}
\item{pval_only}{Label top genes by p-value regardless of candidate status}
\item{candidate_lfc}{Label top significant candidates by fold change magnitude}
}
}
\usage{
label_strategies
}
\description{
Predefined strategies for selecting which genes to label in volcano plots.
}
\keyword{datasets}
