% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_umap_with_hulls}
\alias{plot_umap_with_hulls}
\title{Plot UMAP with cluster hulls}
\usage{
plot_umap_with_hulls(umap_df, alpha = 5, buffer_dist = 0.3)
}
\arguments{
\item{umap_df}{Data frame containing UMAP1, UMAP2, cluster, and value columns}

\item{alpha}{Alpha shape parameter for hull construction (default: 5)}

\item{buffer_dist}{Buffer distance to expand hulls (default: 0.3)}
}
\value{
A tidyplot/ggplot2 object
}
\description{
Plot UMAP with cluster hulls
}
