% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_counts_table}
\alias{process_counts_table}
\title{Process txt output of featureCounts from either a data.table or file path}
\usage{
process_counts_table(x, ...)
}
\arguments{
\item{x}{Either a data.table or character path}
}
\description{
Process txt output of featureCounts from either a data.table or file path
}
